// extension/popup/styles/glassmorphism.module.css.d.ts
declare const styles: {
  container: string;
  glass: string;
  glassCard: string;
  bookmarkItem: string;
  inputWrapper: string;
  glassInput: string;
  glassButton: string;
  glassDisconnect: string;
  statusDot: string;
  connected: string;
  disconnected: string;
  title: string;
  badge: string;
};

export default styles; 